backend/ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# OCR Dependencies
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78
easyocr==1.7.0

# Optional: Cloud OCR services (uncomment if needed)
# google-cloud-vision==3.4.5
# azure-cognitiveservices-vision-computervision==0.9.0
# boto3==1.34.0  # for AWS Textract

# Image processing
numpy==1.26.0
requests==2.31.0